Fixed import (module renamed previously)
